django_project/apps/client/serializers.py
from rest_framework import serializers
from .models import *


class ClientSerializer(serializers.ModelSerializer):
    class Meta:
        model = Client
        fields = ['id', 'name', 'identification', 'sex', 'address', 'photo_url']

