django_project/requirements.txt
asgiref==3.2.3
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
mysqlclient==1.4.6
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.1
